pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "single-house-landscape-planner"
version = "0.1.0"
description = "Landscape planning toolkit for a single-house irregular hillside parcel."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=10.0.0",
]

[project.scripts]
house-landscape = "house_landscape_planner.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
